+++ src/variables/redirects/canvas-catalog/EN.ts
@@ -61,4 +61,4 @@ const EN: RedirectLinkPartial[] = [
   },
 ];
 
-export default EN.map((links) => ({ ...links, lang: "EN" }));
+export default EN.map((links) => ({ ...links, lang: "EN" })) as RedirectLink[];