+++ src/variables/redirects/canvas-lms/DE.ts
@@ -11,4 +11,4 @@ const DE: RedirectLinkPartial[] = [
   },
 ];
 
-export default DE.map((links) => ({ ...links, lang: "DE" }));
+export default DE.map((links) => ({ ...links, lang: "DE" })) as RedirectLink[];