+++ src/variables/redirects/parchment/EN.ts
@@ -1,3 +1,3 @@
 const EN: RedirectLink[] = [];
 
-export default EN.map((links) => ({ ...links, lang: "EN" }));
+export default EN.map((links) => ({ ...links, lang: "EN" })) as RedirectLink[];