+++ CHANGELOG.md
@@ -1,11 +1,19 @@
 # Changelog
 
-## [Unreleased](https://github.com/thedannywahl/instructure-security-package/compare/isp-site-v4.2.1...isp-site)
+## [Unreleased](https://github.com/thedannywahl/instructure-security-package/compare/isp-site-v4.2.2...isp-site)
 
 ### Roadmap
 
 * Add MDX support to markdown renderer
 * Additional contexts (language detection, UI detection, frame detection)
+
+## [4.2.2](https://github.com/thedannywahl/instructure-security-package/compare/isp-site-v4.2.1...isp-site-v4.2.2)
+
+### Fixed
+* Mastery View Formative short link.
+
+### Changed
+* Stricter typing for redirect links.
 
 ## [4.2.1](https://github.com/thedannywahl/instructure-security-package/compare/isp-site-v4.2.0...isp-site-v4.2.1)